useMemo and useCallback Practice
Here i duplicated my app file so i could freely play around with the 2 newest hooks i have been learning about.

To do this i have targeted the Archive component which was re-rendering unnecessarily when using other parts of the application. I first added the Archive to a memo function.

I then memoized the handleAddPost function with useCallback and the archiveOptions with the useMemo hook.

This vastly optimizes rendering of the archive component which is somewhat slow for re-renders, and cuts the amount of wasted re renders.

I also made some optimizations to the context used here. Firstly i passed the components it needs in as children making it so the components are created before being passed in to the context.

I have then wrapped the value we pass to the provider in a useMemo hook thus ensuring it stays persistent through renders unless of course the values in the dependency array change. In this case it still re-renders but that is correctly because the parent re-rendered and not because the context changed eliminating the potential of those wasted re-renders.